Move ServiceIntentionsEntry class and ListIntentions implementation into connect.cs and create a new ServiceIntentionsEntryResponse class to handle Service intention responses
