+++ Consul.Test/ConfigurationTest.cs
@@ -32,12 +32,7 @@ namespace Consul.Test
     [Collection(nameof(ExclusiveCollection))]
     public class ConfigurationTest : BaseFixture
     {
-        private readonly ITestOutputHelper _output;
-
-        public ConfigurationTest(ITestOutputHelper output)
-        {
-            _output = output;
-        }
+      
         [Fact]
         public async Task Configuration_ApplyConfig()
         {
@@ -120,105 +115,6 @@ namespace Consul.Test
             Assert.Null(getDeletedConfigResult.Response);
         }
 
-        [Fact]
-        public async Task Configuration_ListIntentions()
-        {
-            var firstEntry = new ServiceIntentionsEntry
-            {
-                Kind = "service-intentions",
-                Name = "Autobots-Assembler",
-                Sources = new List<SourceIntention>
-                {
-                    new SourceIntention
-                    {
-                        Name = "fortunate",
-                        Action = "allow"
-                    },
-                    new SourceIntention
-                    {
-                        Name = "Prad",
-                        Action = "allow"
-                    }
-                    //new SourceIntention
-                    //{
-                    //    Name = "!Optimus-Prime",
-                    //    Permissions = new List<IntentionPermission>
-                    //    {
-                    //       new IntentionPermission
-                    //       {
-                    //           Action = "allow",
-                    //           HTTP = new IntentionHTTPPermission
-                    //           {
-                    //               PathPrefix = "/",
-                    //               Methods = new List<string>{"GET", "POST"}
-                    //           }
-                    //       }
-                    //    }
-                    //}
-                }
-            };
-
-            var secondEntry = new ServiceIntentionsEntry
-            {
-                Kind = "service-intentions",
-                Name = "*",
-                Sources = new List<SourceIntention>
-                {
-                    new SourceIntention
-                    {
-                        Name = "middle",
-                        Action = "deny"
-                    },
-                    new SourceIntention
-                    {
-                        Name = "center",
-                        Action = "allow"
-                    }
-                    //new SourceIntention
-                    //{
-                    //    Name = "!Megatron",
-                    //    Permissions = new List<IntentionPermission>
-                    //    {
-                    //       new IntentionPermission
-                    //       {
-                    //           Action = "allow",
-                    //           HTTP = new IntentionHTTPPermission
-                    //           {
-                    //               PathPrefix = "/v2",
-                    //               Methods = new List<string>{"DELETE", "POST"}
-                    //           }
-                    //       }
-                    //    }
-                    //}
-                }
-
-            };
-
-
-            var resultOne = await _client.Configuration.ApplyConfig(firstEntry);
-            Assert.Equal(HttpStatusCode.OK, resultOne.StatusCode);
-
-            var resultTwo = await _client.Configuration.ApplyConfig(secondEntry);
-            Assert.Equal(HttpStatusCode.OK, resultTwo.StatusCode);
-
-            var intentionsQuery = await _client.Configuration.ListIntentions<ServiceIntentionsEntry>();
-            Assert.Equal(HttpStatusCode.OK, intentionsQuery.StatusCode);
-
-            var intentions = intentionsQuery.Response;
-            Assert.NotNull(intentions);
-            //foreach (var intention in intentions)
-            //{
-            //    _output.WriteLine($"Intentions: {intention.Name}");
-            //    foreach (var source in intention.Sources)
-            //    {
-            //        _output.WriteLine($"Intentions: {source}");
-            //    }
-            //}
-            Assert.Contains(intentions, i => i.Name == "Autobots-Assembler");
-            Assert.Contains(intentions, i => i.Name == "*");
-
-            await _client.Configuration.DeleteConfig(firstEntry.Kind, firstEntry.Name);
-            await _client.Configuration.DeleteConfig(secondEntry.Kind, secondEntry.Name);
-        }
+      
     }
 }
