+++ Consul/Interfaces/IConnectEndpoint.cs
@@ -32,5 +32,7 @@ namespace Consul.Interfaces
         Task<QueryResult<CAConfig>> CAGetConfig(CancellationToken ct = default);
         Task<WriteResult> CASetConfig(CAConfig config, WriteOptions q, CancellationToken ct = default);
         Task<WriteResult> CASetConfig(CAConfig config, CancellationToken ct = default);
+         Task<QueryResult<List<ServiceIntentionResponse>>> ListIntentions<ServiceIntentionResponse>(CancellationToken ct = default);
+        Task<QueryResult<List<ServiceIntentionResponse>>> ListIntentions<ServiceIntentionResponse>(QueryOptions q, CancellationToken ct = default);
     }
 }
